requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-filter==23.5
python-decouple==3.8
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-cors-headers==4.3.1 
